Move dict utilities into one file
