+++ tavern/request/mqtt.py
@@ -5,8 +5,7 @@ import functools
 from future.utils import raise_from
 
 from tavern.util import exceptions
-from tavern.util.dict_util import format_keys
-from tavern.util.keys import check_expected_keys
+from tavern.util.dict_util import format_keys, check_expected_keys
 
 from .base import BaseRequest
